cirrus-project/tasks/download/requirements.txt
cirrus-geo
stactask
